Disable OpenBLAS and SuperLU.
Somehow, these libraries, when enabled in armadillo making importing
_gdal.so in the gdal wheel very slow.  This slowness has apparently been
occurring since around the time of the release of GDAL 3.1.0.

I don't currently know the underlying cause, as numerous packages have
changed since then, as has the base docker image published for manylinux
wheels.  Using versions of OpenBLAS and SuperLU from that time did not
help, so it is probably libraries they depend on (the versions of gcc
and OpenMP might be the next items to check).

Minor updates.